COM Testing

+++ ara/com/include/types.hpp
@@ -15,8 +15,8 @@ namespace ara
     namespace com
     {
         enum class MethodCallProcessingMode { kPoll, kEvent, kEventSingleThread };
-        using serviceIdentifierType = uint32_t; 
-        using InstanceIdentifier = uint32_t;
+        using serviceIdentifierType = uint16_t; 
+        using InstanceIdentifier = uint16_t;
         class InstanceIdentifierContainer : public std::vector<InstanceIdentifier>{};
         template <typename H>
         class ServiceHandleContainer : public std::vector<H>{};
