+++ ara/com/network_binding/NetworkBinding.cpp
@@ -9,13 +9,13 @@ namespace ara
         /////////////////////////////////////////////////////////////////////////////////////////////////
         ////////////////////////////////////// SOMEIP NETWORK BINDING ///////////////////////////////////
         /////////////////////////////////////////////////////////////////////////////////////////////////
-        SomeIpNetworkBinding::SomeIpNetworkBinding(string ip, int port) : ip{ip}, port{port} {}
+        SomeIpNetworkBinding::SomeIpNetworkBinding(string ip, uint16_t port) : ip{ip}, port{port} {}
 
-        SomeIpNetworkBinding::SomeIpNetworkBinding(int service_id, int instance_id, string ip, int port ,someip::EndUserType type) :
+        SomeIpNetworkBinding::SomeIpNetworkBinding(int service_id, int instance_id, string ip, uint16_t port ,someip::EndUserType type) :
                     ip{ip}, port{port}, serviceId{service_id}, InstanceId{instance_id},someipConfig{someip::TransportProtocol::TCP, type},
                     clientInstance{someip::someipConnection::SetSomeIpConfiguration(io_service, port, someipConfig)} {
+                        
 
-                        ;
                     }
