+++ ara/com/network_binding/SomeIp/servicediscovery/lib/include/servicestorage.hpp
@@ -25,5 +25,5 @@ public:
 static void AddToServiceRegistry(uint16_t Service_ID,serviceinfo serviceinformation); //offer
 static void RemoveService(uint16_t Service_ID,uint16_t instance_id);//stop offer
 static std::vector<serviceinfo> SearchServiceRegistry(uint16_t Service_ID,uint16_t Instance_ID);
-static serviceinfo SetServiceInfo(uint16_t Instance_ID,std::string ipv4_address,uint8_t port_num);
+static serviceinfo SetServiceInfo(uint16_t Instance_ID,std::string ipv4_address,uint16_t port_num);
 };
